Add files via upload
This file contains the code for training and testing AlexNet on T1 post contrast and T2 astrocytomas, oligoastrocytomas, oligodendrogliomas, and normal brain MRIs, AFTER transfer learning from a pretrained camouflage animal detection network. The code for generating feature spaces of the resulting networks is also included in the bottom section.